requirements.txt
fastapi==0.104.1
langchain==0.0.335
pydantic==1.10.0
python-dotenv==1.0.0
SQLAlchemy==2.0.21
uvicorn==0.24.0.post1
